+++ packages/common-cosmos/src/project/models.ts
@@ -21,10 +21,9 @@ import {
   SubqlCosmosMessageHandler,
   CustomModule,
   SubqlCosmosTxFilter,
+  SubqlCosmosProcessorOptions,
 } from '@subql/types-cosmos';
-
 import {plainToClass, Transform, Type} from 'class-transformer';
-
 import {
   IsArray,
   IsEnum,
@@ -34,9 +33,10 @@ import {
   IsObject,
   ValidateNested,
   ValidateIf,
-  IsNumber,
   IsBoolean,
+  Validate,
 } from 'class-validator';
+import {FileReferenceImp} from './utils';
 
 export class CosmosBlockFilter implements SubqlCosmosBlockFilter {
   @IsOptional()
@@ -159,6 +159,12 @@ export class CosmosCustomMapping implements SubqlCosmosMapping<SubqlCosmosCustom
   file: string;
 }
 
+export class CosmosProcessorOptions implements SubqlCosmosProcessorOptions {
+  @IsOptional()
+  @IsString()
+  abi?: string;
+}
+
 export class CosmosRuntimeDataSourceBase<M extends SubqlCosmosMapping<SubqlCosmosRuntimeHandler>>
   implements SubqlCosmosRuntimeDatasource<M>
 {
@@ -172,6 +178,13 @@ export class CosmosRuntimeDataSourceBase<M extends SubqlCosmosMapping<SubqlCosmo
   @Type(() => CosmosCustomModuleImpl)
   @ValidateNested({each: true})
   chainTypes: Map<string, CustomModule>;
+  @IsOptional()
+  @Validate(FileReferenceImp)
+  assets?: Map<string, FileReference>;
+  @IsOptional()
+  @Type(() => CosmosProcessorOptions)
+  @ValidateNested()
+  options?: CosmosProcessorOptions;
 }
 
 export class CosmosFileReferenceImpl implements FileReference {