+++ packages/common-cosmos/src/project/utils.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import {
+  FileReference,
   SecondLayerHandlerProcessor,
   SubqlCosmosCustomDatasource,
   SubqlCosmosDatasource,
@@ -9,6 +10,7 @@ import {
   SubqlCosmosHandlerKind,
   SubqlCosmosRuntimeDatasource,
 } from '@subql/types-cosmos';
+import {ValidationArguments, ValidatorConstraint, ValidatorConstraintInterface} from 'class-validator';
 import {gte} from 'semver';
 import {CustomDatasourceTemplate, RuntimeDatasourceTemplate} from './versioned';
 
@@ -49,4 +51,21 @@ export function isCosmosTemplates(
   specVersion: string
 ): templatesData is (RuntimeDatasourceTemplate | CustomDatasourceTemplate)[] {
   return (isRuntimeCosmosDs(templatesData[0]) || isCustomCosmosDs(templatesData[0])) && gte(specVersion, '0.2.1');
+}
+
+@ValidatorConstraint({name: 'isFileReference', async: false})
+export class FileReferenceImp implements ValidatorConstraintInterface {
+  validate(value: Map<string, FileReference>): boolean {
+    if (!value) {
+      return false;
+    }
+    return !!Object.values(value).find((fileReference: FileReference) => this.isValidFileReference(fileReference));
+  }
+  defaultMessage(args: ValidationArguments): string {
+    return `${JSON.stringify(args.value)} is not a valid assets format`;
+  }
+
+  private isValidFileReference(fileReference: FileReference): boolean {
+    return typeof fileReference === 'object' && 'file' in fileReference && typeof fileReference.file === 'string';
+  }
 }